+++ Sources/Examples/SwiftUI Examples/Testing Examples/SwiftUIRoot.swift
@@ -16,6 +16,10 @@ struct SwiftUIExamples {
             Example("Interactive features", note: "Use featuresets to add interactions to Standard Style.", destination: StandardInteractiveFeaturesExample())
             Example("Interactive buildings", note: "Add interactions to buildings in Standard Style", destination: StandardInteractiveBuildingsExample())
         },
+        Examples.Category("3D") {
+            Example("3D model with feature-state driven interactions", note: "Use expressions and feature state to control model materials and nodes.", destination: Interactive3DModelFeatureStateExample())
+            Example("Animated 3D airplane model", note: "Animate a 3D airplane model along a flight path with animated propellers, landing gear, and lights using feature state.", destination: Animated3DModelSourceExample())
+        },
         Examples.Category("Annotations") {
             Example("Add Map Markers", note: "Add/remove Markers to your map.", destination: MarkersExample())
             Example("View Annotations", note: "Add/remove view annotation on tap.", destination: ViewAnnotationsExample())